+++ dp3t-sdk/sdk/src/main/java/org/dpppt/android/sdk/DP3T.java
@@ -13,6 +13,7 @@ import androidx.core.content.ContextCompat;
 
 import java.io.IOException;
 import java.security.NoSuchAlgorithmException;
+import java.security.PublicKey;
 import java.util.Collection;
 import java.util.Date;
 import java.util.List;
@@ -46,11 +47,11 @@ public class DP3T {
 
 	private static String appId;
 
-	public static void init(Context context, String appId) {
-		init(context, appId, false);
+	public static void init(Context context, String appId, PublicKey signaturePublicKey) {
+		init(context, appId, false, signaturePublicKey);
 	}
 
-	public static void init(Context context, String appId, boolean enableDevDiscoveryMode) {
+	public static void init(Context context, String appId, boolean enableDevDiscoveryMode, PublicKey signaturePublicKey) {
 		if (ProcessUtil.isMainProcess(context)) {
 			DP3T.appId = appId;
 			AppConfigManager appConfigManager = AppConfigManager.getInstance(context);
@@ -58,24 +59,26 @@ public class DP3T {
 			appConfigManager.setDevDiscoveryModeEnabled(enableDevDiscoveryMode);
 			appConfigManager.triggerLoad();
 
-			executeInit(context);
+			executeInit(context, signaturePublicKey);
 		}
 	}
 
-	public static void init(Context context, ApplicationInfo applicationInfo) {
+	public static void init(Context context, ApplicationInfo applicationInfo, PublicKey signaturePublicKey) {
 		if (ProcessUtil.isMainProcess(context)) {
 			DP3T.appId = applicationInfo.getAppId();
 			AppConfigManager appConfigManager = AppConfigManager.getInstance(context);
 			appConfigManager.setManualApplicationInfo(applicationInfo);
 
-			executeInit(context);
+			executeInit(context, signaturePublicKey);
 		}
 	}
 
-	private static void executeInit(Context context) {
+	private static void executeInit(Context context, PublicKey signaturePublicKey) {
 		CryptoModule.getInstance(context).init();
 
 		new Database(context).removeOldData();
+
+		SyncWorker.setBucketSignaturePublicKey(signaturePublicKey);
 
 		AppConfigManager appConfigManager = AppConfigManager.getInstance(context);
 		boolean advertising = appConfigManager.isAdvertisingEnabled();