+++ dp3t-sdk/sdk/src/main/java/org/dpppt/android/sdk/internal/backend/BackendBucketRepository.java
@@ -9,12 +9,14 @@ import android.content.Context;
 import androidx.annotation.NonNull;
 
 import java.io.IOException;
-import java.util.Date;
+import java.security.PublicKey;
 
 import com.google.protobuf.InvalidProtocolBufferException;
 
+import org.dpppt.android.sdk.backend.SignatureVerificationInterceptor;
 import org.dpppt.android.sdk.internal.backend.proto.Exposed;
 
+import io.jsonwebtoken.security.SignatureException;
 import retrofit2.Response;
 import retrofit2.Retrofit;
 import retrofit2.converter.protobuf.ProtoConverterFactory;
@@ -23,14 +25,15 @@ public class BackendBucketRepository implements Repository {
 
 	public static long BATCH_LENGTH = 2 * 60 * 60 * 1000L;
 
-	private static final long ALLOWED_SERVER_TIME_DIFF = 30 * 1000L;
-
 	private BucketService bucketService;
 
-	public BackendBucketRepository(@NonNull Context context, @NonNull String bucketBaseUrl) {
+	public BackendBucketRepository(@NonNull Context context, @NonNull String bucketBaseUrl, @NonNull PublicKey publicKey) {
 		Retrofit bucketRetrofit = new Retrofit.Builder()
 				.baseUrl(bucketBaseUrl)
-				.client(getClientBuilder(context).build())
+				.client(getClientBuilder(context)
+						.addInterceptor(new TimingVerificationInterceptor())
+						.addInterceptor(new SignatureVerificationInterceptor(publicKey))
+						.build())
 				.addConverterFactory(ProtoConverterFactory.create())
 				.build();
 
@@ -38,7 +41,7 @@ public class BackendBucketRepository implements Repository {
 	}
 
 	public Exposed.ProtoExposedList getExposees(long batchReleaseTime)
-			throws IOException, StatusCodeException, ServerTimeOffsetException {
+			throws IOException, StatusCodeException, ServerTimeOffsetException, SignatureException {
 		Response<Exposed.ProtoExposedList> response;
 		try {
 			response = bucketService.getExposees(batchReleaseTime).execute();